Propagate MissingDatasetTypeError to client
Hoist MissingDatasetTypeError up to the top-level exceptions file, and mark it as as ButlerUserError so that it can be sent from the server to the client.  This allows removing some code specific to this exception type from RemoteButler.
