+++ python/lsst/daf/butler/_exceptions_legacy.py
@@ -25,22 +25,27 @@
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-# Re-export some top-level exception types for backwards compatibility -- these
-# used to be part of registry.
-from .._exceptions import MissingDatasetTypeError
-from .._exceptions_legacy import DatasetTypeError, RegistryError
+__all__ = ("DatasetTypeError", "RegistryError")
 
-# Registry imports.
-from . import interfaces, managers, queries, wildcards
-from ._collection_summary import *
-from ._collection_type import *
-from ._config import *
-from ._defaults import *
-from ._exceptions import *
-from ._registry import *
-from ._registry_factory import *
+# The classes in this file exist only for backwards compatibility.  New
+# exception types should not inherit from these.
 
-# Some modules intentionally not imported, either because they are purely
-# internal (e.g. nameShrinker.py) or they contain implementations that are
-# always loaded from configuration strings (e.g. databases subpackage,
-# opaque.py, ...).
+
+class RegistryError(Exception):
+    """Base class for many exception classes produced by Registry methods.
+
+    Notes
+    -----
+    The client code that needs to handle exceptions generated by the Registry
+    methods can catch this class or one of its many subclasses as described by
+    the particular method documentation. While most of the Registry methods
+    should only raise the exceptions of this type, it is hard to guarantee
+    that they will never raise other exception types. If the client needs to
+    handle all possible exceptions, then it should also catch a standard
+    `Exception` type as well. Additionally, some Registry methods can be
+    explicitly documented to raise exceptions outside this class hierarchy.
+    """
+
+
+class DatasetTypeError(RegistryError):
+    """Exception raised for problems with dataset types."""
