+++ python/lsst/daf/butler/registry/_exceptions.py
@@ -34,49 +34,24 @@ __all__ = (
     "ConflictingDefinitionError",
     "DataIdError",
     "DataIdValueError",
-    "DatasetTypeError",
     "DatasetTypeExpressionError",
     "DimensionNameError",
     "InconsistentDataIdError",
     "MissingCollectionError",
-    "MissingDatasetTypeError",
     "MissingSpatialOverlapError",
     "NoDefaultCollectionError",
     "OrphanedRecordError",
     "RegistryConsistencyError",
-    "RegistryError",
     "UnsupportedIdGeneratorError",
     "UserExpressionError",
     "UserExpressionSyntaxError",
 )
 
-
-class RegistryError(Exception):
-    """Base class for many exception classes produced by Registry methods.
-
-    Notes
-    -----
-    The client code that needs to handle exceptions generated by the Registry
-    methods can catch this class or one of its many subclasses as described by
-    the particular method documentation. While most of the Registry methods
-    should only raise the exceptions of this type, it is hard to guarantee
-    that they will never raise other exception types. If the client needs to
-    handle all possible exceptions, then it should also catch a standard
-    `Exception` type as well. Additionally, some Registry methods can be
-    explicitly documented to raise exceptions outside this class hierarchy.
-    """
+from .._exceptions_legacy import RegistryError
 
 
 class ArgumentError(RegistryError):
     """Exception raised when method arguments are invalid or inconsistent."""
-
-
-class DatasetTypeError(RegistryError):
-    """Exception raised for problems with dataset types."""
-
-
-class MissingDatasetTypeError(DatasetTypeError, KeyError):
-    """Exception raised when a dataset type does not exist."""
 
 
 class DatasetTypeExpressionError(RegistryError):
