+++ python/lsst/daf/butler/registry/queries/_query_backend.py
@@ -48,9 +48,10 @@ from lsst.daf.relation import (
 
 from ..._column_tags import DatasetColumnTag, DimensionKeyColumnTag
 from ..._dataset_type import DatasetType
+from ..._exceptions import MissingDatasetTypeError
+from ..._exceptions_legacy import DatasetTypeError
 from ...dimensions import DimensionGroup, DimensionRecordSet, DimensionUniverse
 from .._collection_type import CollectionType
-from .._exceptions import DatasetTypeError, MissingDatasetTypeError
 from ..wildcards import CollectionWildcard
 from ._query_context import QueryContext
 from .find_first_dataset import FindFirstDataset
