+++ python/lsst/daf/butler/registry/tests/_registry.py
@@ -56,6 +56,8 @@ from lsst.daf.relation import Relation, RelationalAlgebraError, Transfer, iterat
 from ..._dataset_association import DatasetAssociation
 from ..._dataset_ref import DatasetIdFactory, DatasetIdGenEnum, DatasetRef
 from ..._dataset_type import DatasetType
+from ..._exceptions import MissingDatasetTypeError
+from ..._exceptions_legacy import DatasetTypeError
 from ..._storage_class import StorageClass
 from ..._timespan import Timespan
 from ...dimensions import DataCoordinate, DataCoordinateSet, SkyPixDimension
@@ -68,11 +70,9 @@ from .._exceptions import (
     CollectionTypeError,
     ConflictingDefinitionError,
     DataIdValueError,
-    DatasetTypeError,
     DatasetTypeExpressionError,
     InconsistentDataIdError,
     MissingCollectionError,
-    MissingDatasetTypeError,
     NoDefaultCollectionError,
     OrphanedRecordError,
 )
